Don't create a useless original index and add an own index for recursive CTEs
